CC-1909: Refactor, move PortMap into it's own file
